Move Test_canvas_Dragged in file with mobile build tag
