+++ internal/driver/mobile/canvas_dragged_test.go
@@ -1,0 +1,39 @@
+//go:build mobile && (!windows || !ci)
+
+package mobile
+
+import (
+	"testing"
+
+	"fyne.io/fyne/v2"
+	"fyne.io/fyne/v2/container"
+	"fyne.io/fyne/v2/widget"
+	"github.com/stretchr/testify/assert"
+)
+
+func Test_canvas_Dragged(t *testing.T) {
+	dragged := false
+	var draggedObj fyne.Draggable
+	scroll := container.NewScroll(widget.NewLabel("Hi\nHi\nHi"))
+	c := newCanvas(fyne.CurrentDevice()).(*canvas)
+	c.SetContent(scroll)
+	c.Resize(fyne.NewSize(40, 24))
+	assert.Equal(t, float32(0), scroll.Offset.Y)
+
+	c.tapDown(fyne.NewPos(32, 3), 0)
+	c.tapMove(fyne.NewPos(32, 10), 0, func(wid fyne.Draggable, ev *fyne.DragEvent) {
+		wid.Dragged(ev)
+		dragged = true
+		draggedObj = wid
+	})
+
+	assert.True(t, dragged)
+	assert.Equal(t, scroll, draggedObj)
+	dragged = false
+	c.tapMove(fyne.NewPos(32, 5), 0, func(wid fyne.Draggable, ev *fyne.DragEvent) {
+		wid.Dragged(ev)
+		dragged = true
+	})
+	assert.True(t, dragged)
+	assert.Equal(t, fyne.NewPos(0, 5), scroll.Offset)
+}
