+++ internal/driver/mobile/canvas_test.go
@@ -75,33 +75,6 @@ func Test_canvas_ChildMinSizeChangeAffectsAncestorsUpToRoot(t *testing.T) {
 	assert.Equal(t, expectedContentSize, content.Size())
 }
 
-func Test_canvas_Dragged(t *testing.T) {
-	dragged := false
-	var draggedObj fyne.Draggable
-	scroll := container.NewScroll(widget.NewLabel("Hi\nHi\nHi"))
-	c := newCanvas(fyne.CurrentDevice()).(*canvas)
-	c.SetContent(scroll)
-	c.Resize(fyne.NewSize(40, 24))
-	assert.Equal(t, float32(0), scroll.Offset.Y)
-
-	c.tapDown(fyne.NewPos(32, 3), 0)
-	c.tapMove(fyne.NewPos(32, 10), 0, func(wid fyne.Draggable, ev *fyne.DragEvent) {
-		wid.Dragged(ev)
-		dragged = true
-		draggedObj = wid
-	})
-
-	assert.True(t, dragged)
-	assert.Equal(t, scroll, draggedObj)
-	dragged = false
-	c.tapMove(fyne.NewPos(32, 5), 0, func(wid fyne.Draggable, ev *fyne.DragEvent) {
-		wid.Dragged(ev)
-		dragged = true
-	})
-	assert.True(t, dragged)
-	assert.Equal(t, fyne.NewPos(0, 5), scroll.Offset)
-}
-
 func Test_canvas_DraggingOutOfWidget(t *testing.T) {
 	c := newCanvas(fyne.CurrentDevice()).(*canvas)
 	slider := widget.NewSlider(0.0, 100.0)
